Add files via upload
In this project, we can search for different types of recipes that is built using HTML, CSS and JavaScript. If the recipe exist in search box then it give the recipe, its ingredients and its cooking method. If any recipe does not exists then it will not give any recipe. We can add our favorite recipes.